rename : rename package to camelCase
패키지명 수정
