+++ app/src/main/AndroidManifest.xml
@@ -19,10 +19,10 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presenter.View.ErrorActivity"
+            android:name=".presenter.view.ErrorActivity"
             android:exported="false" />
         <activity
-            android:name=".presenter.View.MapActivity"
+            android:name=".presenter.view.MapActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -31,7 +31,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presenter.View.PlaceSearchActivity"
+            android:name=".presenter.view.PlaceSearchActivity"
             android:exported="false"></activity>
     </application>
