+++ app/src/main/java/campus/tech/kakao/map/base/ViewModelFactory.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import campus.tech.kakao.map.domain.PlaceRepository
 import campus.tech.kakao.map.presenter.viewModel.MapViewModel
-import campus.tech.kakao.map.ViewModel.SearchViewModel
+import campus.tech.kakao.map.presenter.viewModel.SearchViewModel
 
 class ViewModelFactory(private val repository: PlaceRepository) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
