+++ app/src/main/java/campus/tech/kakao/map/presenter/View/PlaceSearchActivity.kt
@@ -1,95 +1,44 @@
-package campus.tech.kakao.map.presenter.View
+package campus.tech.kakao.map.presenter.view
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.View.GONE
-import android.view.View.VISIBLE
-import android.widget.EditText
-import android.widget.ImageView
 import android.widget.LinearLayout.VERTICAL
-import android.widget.TextView
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.widget.addTextChangedListener
+import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import campus.tech.kakao.map.R
-import campus.tech.kakao.map.domain.VO.Place
-import campus.tech.kakao.map.presenter.View.adapter.FavoriteAdapter
-import campus.tech.kakao.map.presenter.View.adapter.SearchResultAdapter
-import campus.tech.kakao.map.ViewModel.SearchViewModel
+import campus.tech.kakao.map.presenter.view.adapter.FavoriteAdapter
+import campus.tech.kakao.map.presenter.view.adapter.SearchResultAdapter
+import campus.tech.kakao.map.presenter.viewModel.SearchViewModel
+import campus.tech.kakao.map.databinding.ActivityPlaceSearchBinding
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class PlaceSearchActivity : AppCompatActivity() {
     private val viewModel: SearchViewModel by viewModels()
-
-    private lateinit var searchResult: RecyclerView
-    private lateinit var noItem: TextView
-    private lateinit var etSearchPlace: EditText
-    private lateinit var deleteSearch: ImageView
-    private lateinit var favorite: RecyclerView
+    private lateinit var binding: ActivityPlaceSearchBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_place_search)
 
-        searchResult = findViewById<RecyclerView>(R.id.searchResult)
-        etSearchPlace = findViewById<EditText>(R.id.etSearchPlace)
-        noItem = findViewById<TextView>(R.id.noItem)
-        deleteSearch = findViewById<ImageView>(R.id.deleteSearch)
-        favorite = findViewById<RecyclerView>(R.id.favorite)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_place_search)
+        binding.lifecycleOwner = this
+        binding.viewModel = viewModel
 
         settingSearchRecyclerView()
         settingFavoriteRecyclerView()
         setDeleteSearchListener()
-        setEditTextListener()
-
     }
 
     private fun settingSearchRecyclerView() {
         setSearchAdapter()
-        searchResult.layoutManager = LinearLayoutManager(this)
-        searchResult.addItemDecoration(
+        binding.searchResult.layoutManager = LinearLayoutManager(this)
+        binding.searchResult.addItemDecoration(
             DividerItemDecoration(this, VERTICAL)
         )
-    }
-
-    private fun settingFavoriteRecyclerView() {
-        setFavoriteAdapter()
-        favorite.layoutManager = LinearLayoutManager(this, LinearLayoutManager.HORIZONTAL, true)
-    }
-
-    private fun setFavoriteAdapter() {
-        val adapter = FavoriteAdapter(
-            onClickDelete = {
-                viewModel.deleteFromFavorite(it)
-            })
-
-        favorite.adapter = adapter
-        viewModel.favoritePlace.observe(this) {
-            adapter.submitList(it)
-            favorite.smoothScrollToPosition(maxOf(it.size-1,0))
-        }
-    }
-
-
-    private fun setDeleteSearchListener() {
-        deleteSearch.setOnClickListener {
-            etSearchPlace.setText("")
-        }
-    }
-
-    private fun setEditTextListener() {
-        etSearchPlace.addTextChangedListener {
-            CoroutineScope(Dispatchers.IO).launch {
-                viewModel.searchPlaceRemote(etSearchPlace.text.toString())
-            }
-        }
     }
 
     private fun setSearchAdapter() {
@@ -100,9 +49,33 @@ class PlaceSearchActivity : AppCompatActivity() {
             })
         viewModel.currentResult.observe(this) {
             adapter.submitList(it)
-            handleVisibility(it)
         }
-        searchResult.adapter = adapter
+        binding.searchResult.adapter = adapter
+    }
+
+    private fun settingFavoriteRecyclerView() {
+        setFavoriteAdapter()
+        binding.favorite.layoutManager = LinearLayoutManager(this, LinearLayoutManager.HORIZONTAL, true)
+    }
+
+    private fun setFavoriteAdapter() {
+        val adapter = FavoriteAdapter(
+            onClickDelete = {
+                viewModel.deleteFromFavorite(it)
+            })
+
+        binding.favorite.adapter = adapter
+        viewModel.favoritePlace.observe(this) {
+            adapter.submitList(it)
+            binding.favorite.smoothScrollToPosition(maxOf(it.size-1,0))
+        }
+    }
+
+
+    private fun setDeleteSearchListener() {
+        binding.deleteSearch.setOnClickListener {
+            binding.etSearchPlace.setText("")
+        }
     }
 
     private fun moveToMapActivity(id:Int){
@@ -110,16 +83,6 @@ class PlaceSearchActivity : AppCompatActivity() {
         intent.putExtra(INTENT_ID,id)
         setResult(RESULT_OK,intent)
         finish()
-    }
-
-    private fun handleVisibility(places : List<Place>){
-        if(places.isEmpty()){
-            searchResult.visibility = GONE
-            noItem.visibility = VISIBLE
-        } else {
-            searchResult.visibility = VISIBLE
-            noItem.visibility = GONE
-        }
     }
 
     companion object{
