+++ app/src/main/java/campus/tech/kakao/map/presenter/viewModel/SearchViewModel.kt
@@ -1,10 +1,10 @@
-package campus.tech.kakao.map.ViewModel
+package campus.tech.kakao.map.presenter.viewModel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import campus.tech.kakao.map.domain.VO.Place
+import campus.tech.kakao.map.domain.vo.Place
 import campus.tech.kakao.map.domain.PlaceRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -31,8 +31,10 @@ class SearchViewModel @Inject constructor(
         _currentResult.postValue(repository.getSimilarPlacesByName(string))
     }
 
-    suspend fun searchPlaceRemote(name: String) {
-        _currentResult.postValue(repository.searchPlaceRemote(name))
+    fun searchPlaceRemote(name: String) {
+        viewModelScope.launch {
+            _currentResult.postValue(repository.searchPlaceRemote(name))
+        }
     }
 
     fun addFavorite(id : Int) {
