+++ app/src/main/res/layout/activity_error.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presenter.View.ErrorActivity">
+    tools:context=".presenter.view.ErrorActivity">
 
     <TextView
         android:id="@+id/errorText"
