+++ app/src/main/res/layout/activity_map.xml
@@ -11,7 +11,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presenter.View.MapActivity">
+        tools:context=".presenter.view.MapActivity">
 
 
         <TextView
