+++ app/src/main/res/layout/place_element.xml
@@ -1,46 +1,54 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="70dp"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <ImageView
-        android:id="@+id/placeIcon"
-        android:layout_width="50dp"
-        android:layout_height="50dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_marginHorizontal="10dp"
-        android:background="@drawable/place_image"/>
-
-    <TextView
-        android:id="@+id/placeName"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        tools:text="카페1"
-        android:textSize="20dp"
-        app:layout_constraintStart_toEndOf="@id/placeIcon"
-        app:layout_constraintTop_toTopOf="@+id/placeIcon"
-        />
-
-    <TextView
-        android:id="@+id/placeAddress"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        tools:text="서울 성동구 성수동 1"
-        android:textColor="@color/grey"
-        app:layout_constraintStart_toEndOf="@id/placeIcon"
-        app:layout_constraintTop_toBottomOf="@+id/placeName"/>
-
-    <TextView
-        android:id="@+id/placeCategory"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        tools:text="카페"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/placeIcon"
-        android:layout_marginRight="10dp"
-        android:textColor="@color/grey"/>
-</androidx.constraintlayout.widget.ConstraintLayout>
+    <data>
+        <variable
+            name="placeItem"
+            type="campus.tech.kakao.map.domain.vo.Place" />
+    </data>
+    <androidx.constraintlayout.widget.ConstraintLayout 
+        android:layout_width="match_parent"
+        android:layout_height="70dp">
+    
+        <ImageView
+            android:id="@+id/placeIcon"
+            android:layout_width="50dp"
+            android:layout_height="50dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_marginHorizontal="10dp"
+            android:background="@drawable/place_image"/>
+    
+        <TextView
+            android:id="@+id/placeName"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@{placeItem.name}"
+            android:textSize="20dp"
+            app:layout_constraintStart_toEndOf="@id/placeIcon"
+            app:layout_constraintTop_toTopOf="@+id/placeIcon"
+            />
+    
+        <TextView
+            android:id="@+id/placeAddress"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@{placeItem.address}"
+            android:textColor="@color/grey"
+            app:layout_constraintStart_toEndOf="@id/placeIcon"
+            app:layout_constraintTop_toBottomOf="@+id/placeName"/>
+    
+        <TextView
+            android:id="@+id/placeCategory"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            tools:text="카페"
+            android:text="@{placeItem.category.toString()}"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="@+id/placeIcon"
+            android:layout_marginRight="10dp"
+            android:textColor="@color/grey"/>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
